+++ team1-api/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     implementation project(':team1-domain')
     implementation project(':team1-common')
 
+    implementation 'org.json:json:20240303'
     //swagger
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.3'
