+++ team1-api/src/main/java/goorm/deepdive/team1/api/kakao/KakaoApiAddressService.java
@@ -1,4 +1,4 @@
-package goorm.deepdive.team1.domain.kakao;
+package goorm.deepdive.team1.api.kakao;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
@@ -8,7 +8,7 @@ import java.net.URL;
 
 import goorm.deepdive.team1.common.exception.CustomException;
 import goorm.deepdive.team1.common.exception.KakaoApiExceptionCode;
-import goorm.deepdive.team1.domain.kakao.Dto.AddressResponseDto;
+import goorm.deepdive.team1.api.kakao.response.AddressResponseDto;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.stereotype.Service;
@@ -16,21 +16,9 @@ import org.springframework.stereotype.Service;
 @Service
 public class KakaoApiAddressService {
 
-    private static final String REST_API_KEY = "9cd742e01ee31be1eb4ac4d5fb33500f";
+    private final String REST_API_KEY = "9cd742e01ee31be1eb4ac4d5fb33500f";
 
-    public static void main(String[] args) {
-        String address = "경기도 고양시 덕양구 화중로 219"; // 🔹 변환하고 싶은 주소
-        AddressResponseDto response = getGeoDataFromAddress(address);
-
-            System.out.println("지번 주소: " + response.getRegionAddress());
-            System.out.println("도로명 주소: " + response.getRoadAddress());
-            System.out.println("위도: " + response.getY());
-            System.out.println("경도: " + response.getX());
-
-    }
-
-
-    public static AddressResponseDto getGeoDataFromAddress(String address) {
+    public AddressResponseDto getGeoDataFromAddress(String address) {
         try {
             String apiUrl = "https://dapi.kakao.com/v2/local/search/address.json";
             String encodedAddress = URLEncoder.encode(address, "UTF-8");
@@ -43,7 +31,7 @@ public class KakaoApiAddressService {
             conn.setRequestProperty("Content-Type", "application/json;charset=UTF-8");
 
             int responseCode = conn.getResponseCode();
-            System.out.println("responseCode = " + responseCode);
+
             if (responseCode == 200) {
                 BufferedReader br = new BufferedReader(new InputStreamReader(conn.getInputStream(), "UTF-8"));
                 StringBuilder response = new StringBuilder();
@@ -66,7 +54,7 @@ public class KakaoApiAddressService {
         }
     }
 
-    private static CustomException getKakaoApiException(int responseCode) {
+    private CustomException getKakaoApiException(int responseCode) {
         return switch (responseCode) {
             case 400 -> new CustomException(KakaoApiExceptionCode.INVALID_REQUEST);
             case 403 -> new CustomException(KakaoApiExceptionCode.UNAUTHORIZED);
@@ -76,7 +64,7 @@ public class KakaoApiAddressService {
         };
     }
 
-    private static AddressResponseDto parseJsonResponse(String jsonResponse) {
+    private AddressResponseDto parseJsonResponse(String jsonResponse) {
         JSONObject jsonObject = new JSONObject(jsonResponse);
         JSONArray documents = jsonObject.getJSONArray("documents");
 
@@ -86,8 +74,8 @@ public class KakaoApiAddressService {
 
         JSONObject data = documents.getJSONObject(0);
 
-        double latitude = data.getDouble("y");
-        double longitude = data.getDouble("x");
+        double latitude = data.getDouble("y");  //위도
+        double longitude = data.getDouble("x");  //경도
 
         JSONObject address = data.optJSONObject("address");
         JSONObject roadAddress = data.optJSONObject("road_address");
@@ -104,8 +92,8 @@ public class KakaoApiAddressService {
         }
 
         return AddressResponseDto.builder()
-                .x(latitude)
-                .y(longitude)
+                .y(latitude)
+                .x(longitude)
                 .regionAddress(region)
                 .roadAddress(roadName)
                 .build();
